+++ changelog.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.7.0 (2025-09-19)
+------------------
+
+* When serializing to HAL, it's now possible to specify that certain
+  relationships should be in _embedded.
+
+
 0.6.1 (2025-09-12)
 ------------------